+++ src/constants.ts
@@ -1,13 +1,13 @@
 import { Group } from './types';
 
-export const COMMIT_FORMAT_PREFIX = /^(break|build|ci|deps|docs|feature|fix|internal|misc|new|patch|release|revert|security|style|test|update)(?:\(([a-zA-Z0-9\-., ]+)\))?:/u;
+export const COMMIT_FORMAT_PREFIX = /^(break|breaking|build|ci|deps|docs|feature|fix|internal|misc|new|patch|release|revert|security|style|test|update)(?:\(([a-zA-Z0-9\-., ]+)\))?:/u;
 
 export const GROUPS: Group[] = [
   {
     bump: 'major',
     emoji: '💥',
     label: 'Breaking',
-    types: ['break'],
+    types: ['break', 'breaking'],
   },
   {
     bump: 'patch',